--- Dishify2/app/Procfile
@@ -1,1 +1,0 @@
-web: gunicorn -w 4 -k uvicorn.workers.UvicornnWorker application:app
