+++ Dishify2/app/__init__.py
@@ -1,0 +1,8 @@
+'''
+Entry point for the app
+'''
+
+from app.application import create_app
+
+APP = create_app()
+APP.run(debug=True)
