+++ Dishify2/app/requirements.txt
@@ -11,9 +11,7 @@ grpcio==1.28.1
 h11==0.9.0
 h5py==2.10.0
 idna==2.9
-Keras==2.3.1
-Keras-Applications==1.0.8
-Keras-Preprocessing==1.1.0
+gunicorn
 numpy==1.18.4
 protobuf==3.11.3
 psycopg2==2.8.5
