+++ Procfile
@@ -1,0 +1,1 @@
+web: gunicorn Dishify2.app:APP -t 120
